+++ tests/test_version.py
@@ -1,9 +1,13 @@
+# encoding: utf-8
+
 from __future__ import unicode_literals
 
 from distutils.version import StrictVersion as SV
 import unittest
 
 import spotify
+
+from tests import mock
 
 
 class VersionTest(unittest.TestCase):
@@ -13,3 +17,22 @@ class VersionTest(unittest.TestCase):
 
     def test_version_is_grater_than_all_1_x_versions(self):
         self.assertLess(SV('1.999'), SV(spotify.__version__))
+
+
+@mock.patch('spotify.version.lib', spec=spotify.lib)
+class LibspotifyVersionTest(unittest.TestCase):
+
+    def test_libspotify_api_version(self, lib_mock):
+        lib_mock.SPOTIFY_API_VERSION = 73
+
+        result = spotify.get_libspotify_api_version()
+
+        self.assertEqual(result, 73)
+
+    def test_libspotify_build_id(self, lib_mock):
+        build_id = spotify.ffi.new('char []', b'12.1.51.foobaræøå')
+        lib_mock.sp_build_id.return_value = build_id
+
+        result = spotify.get_libspotify_build_id()
+
+        self.assertEqual(result, '12.1.51.foobaræøå')